requirements.txt
openai>=1.0.0
langchain>=0.2.0
langchain-openai>=0.1.0
langgraph>=0.2.0
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.26.0
psycopg2-binary>=2.9.0
pgvector>=0.2.0
sqlalchemy>=2.0.0
alembic>=1.13.0
Pillow>=10.0.0
python-dotenv>=1.0.0
